Logipool Website
We designed and developed a website using angular. The website contains all the information required for the student to enroll the courses. We provided all the courses details using dummy database as well as we provided the facility to register for various courses using registration forms, we have used forms validations in the form, we also added a contact page through which students can easily reach out to the institute.